+++ example/lib/page_structure.dart
@@ -10,17 +10,15 @@ import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 import 'package:flutter_zoom_drawer/flutter_zoom_drawer.dart';
 import 'package:provider/provider.dart';
 
-
-
 class PageStructure extends StatelessWidget {
-  final String title;
-  final Widget child;
-  final List<Widget> actions;
-  final Color backgroundColor;
-  final double elevation;
+  final String? title;
+  final Widget? child;
+  final List<Widget>? actions;
+  final Color? backgroundColor;
+  final double? elevation;
 
   const PageStructure({
-    Key key,
+    Key? key,
     this.title,
     this.child,
     this.actions,
@@ -30,7 +28,7 @@ class PageStructure extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final angle = true ? 180 * pi / 180 : 0.0;
+    final angle = ZoomDrawer.isRTL() ? 180 * pi / 180 : 0.0;
     final _currentPage =
         context.select<MenuProvider, int>((provider) => provider.currentPage);
     final container = Container(
@@ -49,7 +47,7 @@ class PageStructure extends StatelessWidget {
       backgroundColor: Colors.transparent,
       appBar: PlatformAppBar(
         automaticallyImplyLeading: false,
-        android: (_) => MaterialAppBarData(elevation: elevation),
+        material: (_, __) => MaterialAppBarData(elevation: elevation),
         title: PlatformText(
           HomeScreen.mainMenu[_currentPage].title,
         ),
@@ -60,10 +58,7 @@ class PageStructure extends StatelessWidget {
               Icons.menu,
             ),
             onPressed: () {
-              // ZoomDrawer.of(context).toggle();
-              AnimatedDrawerWidget.of(context).toggle();
-              // RenderDefault.of(context).toggle();
-
+              ZoomDrawer.of(context)!.toggle();
             },
           ),
         ),