+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# [2.0.0]
+
+* Migrated to Null Safety
+* Added a parameter to disable the swipe gesture (`disableGesture`)
+
+
+# [1.0.5]
+
+* Added multiple drawer styles thanks to @dangtienngoc
+
 # [1.0.4+1]
 
 * Fixed animation controller dispose issue